style: :lipstick: Make separate components for `Member` and `Team`. Adjust font and social link colors.
